+++ src/cultionet/data/create.py
@@ -1,181 +1,36 @@
 import typing as T
-from functools import partial
 from pathlib import Path
 
-import cv2
+import dask
 import einops
 import geopandas as gpd
 import geowombat as gw
-import joblib
 import numpy as np
 import pandas as pd
 import torch
 import xarray as xr
 from affine import Affine
-from geowombat.core.windows import get_window_offsets
-from joblib import Parallel, delayed, parallel_backend
-from rasterio.dtypes import get_minimum_dtype
-from rasterio.features import rasterize as rio_rasterize
-from rasterio.warp import calculate_default_transform
-from rasterio.windows import Window, from_bounds
+from dask.distributed import Client, LocalCluster, progress
+from rasterio.windows import from_bounds
 from scipy.ndimage import label as nd_label
-from scipy.ndimage import uniform_filter
 from skimage.measure import regionprops
 from threadpoolctl import threadpool_limits
 
 from ..augment.augmenters import AugmenterMapping
 from ..utils.logging import set_color_logger
 from .data import Data, LabeledData
-from .utils import get_image_list_dims
+from .store import BatchStore
+from .utils import (
+    cleanup_edges,
+    edge_gradient,
+    fillz,
+    get_crop_count,
+    get_image_list_dims,
+    normalize_boundary_distances,
+    polygon_to_array,
+)
 
 logger = set_color_logger(__name__)
-
-
-def roll(
-    arr_pad: np.ndarray,
-    shift: T.Union[int, T.Tuple[int, int]],
-    axis: T.Union[int, T.Tuple[int, int]],
-) -> np.ndarray:
-    """Rolls array elements along a given axis and slices off padded edges."""
-    return np.roll(arr_pad, shift, axis=axis)[1:-1, 1:-1]
-
-
-def close_edge_ends(array: np.ndarray) -> np.ndarray:
-    """Closes 1 pixel gaps at image edges."""
-    # Top
-    idx = np.where(array[1] == 1)
-    z = np.zeros(array.shape[1], dtype="uint8")
-    z[idx] = 1
-    array[0] = z
-    # Bottom
-    idx = np.where(array[-2] == 1)
-    z = np.zeros(array.shape[1], dtype="uint8")
-    z[idx] = 1
-    array[-1] = z
-    # Left
-    idx = np.where(array[:, 1] == 1)
-    z = np.zeros(array.shape[0], dtype="uint8")
-    z[idx] = 1
-    array[:, 0] = z
-    # Right
-    idx = np.where(array[:, -2] == 1)
-    z = np.zeros(array.shape[0], dtype="uint8")
-    z[idx] = 1
-    array[:, -1] = z
-
-    return array
-
-
-def get_other_crop_count(array: np.ndarray) -> np.ndarray:
-    array_pad = np.pad(array, pad_width=((1, 1), (1, 1)), mode="edge")
-
-    rarray = roll(array_pad, 1, axis=0)
-    crop_count = np.uint8((rarray > 0) & (rarray != array) & (array > 0))
-    rarray = roll(array_pad, -1, axis=0)
-    crop_count += np.uint8((rarray > 0) & (rarray != array) & (array > 0))
-    rarray = roll(array_pad, 1, axis=1)
-    crop_count += np.uint8((rarray > 0) & (rarray != array) & (array > 0))
-    rarray = roll(array_pad, -1, axis=1)
-    crop_count += np.uint8((rarray > 0) & (rarray != array) & (array > 0))
-
-    return crop_count
-
-
-def fill_edge_gaps(labels: np.ndarray, array: np.ndarray) -> np.ndarray:
-    """Fills neighboring 1-pixel edge gaps."""
-    # array_pad = np.pad(array, pad_width=((1, 1), (1, 1)), mode='edge')
-    # hsum = roll(array_pad, 1, axis=0) + roll(array_pad, -1, axis=0)
-    # vsum = roll(array_pad, 1, axis=1) + roll(array_pad, -1, axis=1)
-    # array = np.where(
-    #     (hsum == 2) & (vsum == 0), 1, array
-    # )
-    # array = np.where(
-    #     (hsum == 0) & (vsum == 2), 1, array
-    # )
-    other_count = get_other_crop_count(np.where(array == 1, 0, labels))
-    array = np.where(other_count > 0, 1, array)
-
-    return array
-
-
-def get_crop_count(array: np.ndarray, edge_class: int) -> np.ndarray:
-    array_pad = np.pad(array, pad_width=((1, 1), (1, 1)), mode="edge")
-
-    rarray = roll(array_pad, 1, axis=0)
-    crop_count = np.uint8((rarray > 0) & (rarray != edge_class))
-    rarray = roll(array_pad, -1, axis=0)
-    crop_count += np.uint8((rarray > 0) & (rarray != edge_class))
-    rarray = roll(array_pad, 1, axis=1)
-    crop_count += np.uint8((rarray > 0) & (rarray != edge_class))
-    rarray = roll(array_pad, -1, axis=1)
-    crop_count += np.uint8((rarray > 0) & (rarray != edge_class))
-
-    return crop_count
-
-
-def get_edge_count(array: np.ndarray, edge_class: int) -> np.ndarray:
-    array_pad = np.pad(array, pad_width=((1, 1), (1, 1)), mode="edge")
-
-    edge_count = np.uint8(roll(array_pad, 1, axis=0) == edge_class)
-    edge_count += np.uint8(roll(array_pad, -1, axis=0) == edge_class)
-    edge_count += np.uint8(roll(array_pad, 1, axis=1) == edge_class)
-    edge_count += np.uint8(roll(array_pad, -1, axis=1) == edge_class)
-
-    return edge_count
-
-
-def get_non_count(array: np.ndarray) -> np.ndarray:
-    array_pad = np.pad(array, pad_width=((1, 1), (1, 1)), mode="edge")
-
-    non_count = np.uint8(roll(array_pad, 1, axis=0) == 0)
-    non_count += np.uint8(roll(array_pad, -1, axis=0) == 0)
-    non_count += np.uint8(roll(array_pad, 1, axis=1) == 0)
-    non_count += np.uint8(roll(array_pad, -1, axis=1) == 0)
-
-    return non_count
-
-
-def cleanup_edges(
-    array: np.ndarray,
-    original: np.ndarray,
-    edge_class: int,
-) -> np.ndarray:
-    """Removes crop pixels that border non-crop pixels."""
-    array_pad = np.pad(original, pad_width=((1, 1), (1, 1)), mode="edge")
-    original_zero = np.uint8(roll(array_pad, 1, axis=0) == 0)
-    original_zero += np.uint8(roll(array_pad, -1, axis=0) == 0)
-    original_zero += np.uint8(roll(array_pad, 1, axis=1) == 0)
-    original_zero += np.uint8(roll(array_pad, -1, axis=1) == 0)
-
-    # Fill edges
-    array = np.where(
-        (array == 0)
-        & (get_crop_count(array, edge_class) > 0)
-        & (get_edge_count(array, edge_class) > 0),
-        edge_class,
-        array,
-    )
-    # Remove crops next to non-crop
-    array = np.where(
-        (array > 0)
-        & (array != edge_class)
-        & (get_non_count(array) > 0)
-        & (get_edge_count(array, edge_class) > 0),
-        0,
-        array,
-    )
-    # Fill in non-cropland
-    array = np.where(original_zero == 4, 0, array)
-    # Remove isolated crop pixels (i.e., crop clumps with 2 or fewer pixels)
-    array = np.where(
-        (array > 0)
-        & (array != edge_class)
-        & (get_crop_count(array, edge_class) <= 1),
-        0,
-        array,
-    )
-
-    return array
 
 
 def is_grid_processed(
@@ -209,117 +64,120 @@ def is_grid_processed(
     return all(batches_stored)
 
 
-def create_boundary_distances(
-    labels_array: np.ndarray, train_type: str, cell_res: float
-) -> T.Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
-    """Creates distances from boundaries."""
-    if train_type.lower() == "polygon":
-        mask = np.uint8(labels_array)
-    else:
-        mask = np.uint8(1 - labels_array)
-    # Get unique segments
-    segments = nd_label(mask)[0]
-    # Get the distance from edges
-    bdist = cv2.distanceTransform(mask, cv2.DIST_L2, 3)
-    bdist *= cell_res
+def reshape_and_mask_array(
+    data: xr.DataArray,
+    num_time: int,
+    num_bands: int,
+    gain: float,
+    offset: int,
+) -> xr.DataArray:
+    """Reshapes an array and masks no-data values."""
 
-    grad_x = cv2.Sobel(
-        np.pad(bdist, 5, mode="edge"), cv2.CV_32F, dx=1, dy=0, ksize=5
-    )
-    grad_y = cv2.Sobel(
-        np.pad(bdist, 5, mode="edge"), cv2.CV_32F, dx=0, dy=1, ksize=5
-    )
-    ori = cv2.phase(grad_x, grad_y, angleInDegrees=False)
-    ori = ori[5:-5, 5:-5] / np.deg2rad(360)
-    ori[labels_array == 0] = 0
-
-    return mask, segments, bdist, ori
-
-
-def normalize_boundary_distances(
-    labels_array: np.ndarray,
-    train_type: str,
-    cell_res: float,
-    normalize: bool = True,
-) -> T.Tuple[np.ndarray, np.ndarray]:
-    """Normalizes boundary distances."""
-
-    # Create the boundary distances
-    __, segments, bdist, ori = create_boundary_distances(
-        labels_array, train_type, cell_res
-    )
-    dist_max = 1e9
-    if normalize:
-        dist_max = 1.0
-        # Normalize each segment by the local max distance
-        props = regionprops(segments, intensity_image=bdist)
-        for p in props:
-            if p.label > 0:
-                bdist = np.where(
-                    segments == p.label, bdist / p.max_intensity, bdist
-                )
-    bdist = np.nan_to_num(
-        bdist.clip(0, dist_max), nan=1.0, neginf=1.0, posinf=1.0
-    )
-    ori = np.nan_to_num(ori.clip(0, 1), nan=1.0, neginf=1.0, posinf=1.0)
-
-    return bdist, ori
-
-
-def edge_gradient(array: np.ndarray) -> np.ndarray:
-    """Calculates the morphological gradient of crop fields."""
-    se = np.array([[1, 1], [1, 1]], dtype="uint8")
-    array = np.uint8(
-        cv2.morphologyEx(np.uint8(array), cv2.MORPH_GRADIENT, se) > 0
+    src_ts_stack = xr.DataArray(
+        # Date are stored [(band x time) x height x width]
+        (
+            data.data.reshape(
+                num_bands,
+                num_time,
+                data.gw.nrows,
+                data.gw.ncols,
+            ).transpose(1, 0, 2, 3)
+        ).astype('float32'),
+        dims=('time', 'band', 'y', 'x'),
+        coords={
+            'time': range(num_time),
+            'band': range(num_bands),
+            'y': data.y,
+            'x': data.x,
+        },
+        attrs=data.attrs.copy(),
     )
 
-    return array
+    with xr.set_options(keep_attrs=True):
+        time_series = (src_ts_stack.gw.mask_nodata() * gain + offset).fillna(0)
+
+    return time_series
 
 
-def polygon_to_array(
-    df: gpd.GeoDataFrame,
-    reference_data: xr.DataArray,
-    column: str,
-    fill_value: int = 0,
-    default_value: int = 1,
-    all_touched: bool = False,
-    dtype: str = "uint8",
-) -> np.ndarray:
-    """Converts a polygon, or polygons, to an array."""
+def create_predict_dataset(
+    image_list: T.List[T.List[T.Union[str, Path]]],
+    region: str,
+    process_path: Path = None,
+    date_format: str = "%Y%j",
+    gain: float = 1e-4,
+    offset: float = 0.0,
+    ref_res: T.Union[float, T.Tuple[float, float]] = 10.0,
+    resampling: str = "nearest",
+    window_size: int = 100,
+    padding: int = 101,
+    num_workers: int = 1,
+    compress_method: T.Union[int, str] = 'zlib',
+):
+    """Creates a prediction dataset for an image."""
 
-    df = df.copy()
+    with gw.config.update(ref_res=ref_res):
+        with gw.open(
+            image_list,
+            stack_dim="band",
+            band_names=list(range(1, len(image_list) + 1)),
+            resampling=resampling,
+            chunks=512,
+        ) as src_ts:
 
-    if df.crs != reference_data.crs:
-        # Transform the geometry
-        df = df.to_crs(reference_data.crs)
+            num_time, num_bands = get_image_list_dims(image_list, src_ts)
 
-    # Get the reference bounds
-    left, bottom, right, top = reference_data.gw.bounds
-    # Get intersecting polygons
-    df = df.cx[left:right, bottom:top]
-    # Clip the polygons to the reference bounds
-    df = gpd.clip(df, reference_data.gw.geodataframe)
+            time_series: xr.DataArray = reshape_and_mask_array(
+                data=src_ts,
+                num_time=num_time,
+                num_bands=num_bands,
+                gain=gain,
+                offset=offset,
+            )
 
-    # Get the output dimensions
-    dst_transform = Affine(
-        reference_data.gw.cellx, 0.0, left, 0.0, -reference_data.gw.celly, top
-    )
+            # Chunk the array into the windows
+            time_series_array = time_series.chunk(
+                {"time": -1, "band": -1, "y": window_size, "x": window_size}
+            ).data.map_overlap(
+                lambda x: x,
+                depth={0: 0, 1: 0, 2: padding, 3: padding},
+                boundary=0,
+                trim=False,
+            )
 
-    # Get the shape geometry and encoding value
-    shapes = list(zip(df.geometry, df[column]))
-
-    # Convert the polygon(s) to an array
-    polygon_array = rio_rasterize(
-        shapes,
-        out_shape=(reference_data.gw.nrows, reference_data.gw.ncols),
-        fill=fill_value,
-        transform=dst_transform,
-        all_touched=all_touched,
-        default_value=default_value,
-        dtype=dtype,
-    )
-
-    return polygon_array
+            with dask.config.set(
+                {
+                    "distributed.worker.memory.terminate": False,
+                    "distributed.comm.retry.count": 10,
+                    "distributed.comm.timeouts.connect": 30,
+                }
+            ):
+                with LocalCluster(
+                    processes=True,
+                    n_workers=num_workers,
+                    threads_per_worker=2,
+                    memory_limit="4GB",  # per worker limit
+                ) as cluster:
+                    with Client(cluster) as client:
+                        with BatchStore(
+                            data=time_series,
+                            write_path=process_path,
+                            res=ref_res,
+                            resampling=resampling,
+                            region=region,
+                            start_date=pd.to_datetime(
+                                Path(image_list[0]).stem, format=date_format
+                            ).strftime("%Y%m%d"),
+                            end_date=pd.to_datetime(
+                                Path(image_list[-1]).stem, format=date_format
+                            ).strftime("%Y%m%d"),
+                            window_size=window_size,
+                            padding=padding,
+                            compress_method=compress_method,
+                            gain=gain,
+                        ) as batch_store:
+                            save_tasks = batch_store.save(time_series_array)
+                            results = client.persist(save_tasks)
+                            progress(results)
 
 
 class ReferenceArrays:
@@ -437,49 +295,6 @@ class ReferenceArrays:
             orientation=orientation,
             edge_array=edge_array,
         )
-
-
-def reshape_and_mask_array(
-    data: xr.DataArray,
-    num_time: int,
-    num_bands: int,
-    gain: float,
-    offset: int,
-) -> xr.DataArray:
-    """Reshapes an array and masks no-data values."""
-
-    src_ts_stack = xr.DataArray(
-        # Date are stored [(band x time) x height x width]
-        (
-            data.data.reshape(
-                num_bands,
-                num_time,
-                data.gw.nrows,
-                data.gw.ncols,
-            ).transpose(1, 0, 2, 3)
-        ).astype('float32'),
-        dims=('time', 'band', 'y', 'x'),
-        coords={
-            'time': range(num_time),
-            'band': range(num_bands),
-            'y': data.y,
-            'x': data.x,
-        },
-        attrs=data.attrs.copy(),
-    )
-
-    with xr.set_options(keep_attrs=True):
-        time_series = (src_ts_stack.gw.mask_nodata() * gain + offset).fillna(0)
-
-    return time_series
-
-
-def fillz(x: np.ndarray) -> np.ndarray:
-    """Fills zeros with the focal mean value."""
-
-    focal_mean = uniform_filter(x, size=(0, 0, 3, 3), mode='reflect')
-
-    return np.where(x == 0, focal_mean, x)
 
 
 class ImageVariables:
@@ -671,230 +486,6 @@ class ImageVariables:
             num_time=num_time,
             num_bands=num_bands,
         )
-
-
-def save_and_update(
-    write_path: Path, predict_data: Data, name: str, compress: int = 5
-) -> None:
-    predict_path = write_path / f"data_{name}.pt"
-    joblib.dump(predict_data, predict_path, compress=compress)
-
-
-def read_slice(darray: xr.DataArray, w_pad: Window) -> xr.DataArray:
-    slicer = (
-        slice(0, None),
-        slice(0, None),
-        slice(w_pad.row_off, w_pad.row_off + w_pad.height),
-        slice(w_pad.col_off, w_pad.col_off + w_pad.width),
-    )
-
-    return darray[slicer]
-
-
-def get_window_chunk(windows: T.List[T.Tuple[Window, Window]], chunksize: int):
-    for i in range(0, len(windows), chunksize):
-        yield windows[i : i + chunksize]
-
-
-def create_and_save_window(
-    write_path: Path,
-    res: float,
-    resampling: str,
-    region: str,
-    start_date: str,
-    end_date: str,
-    window_size: int,
-    padding: int,
-    compress_method: T.Union[int, str],
-    darray: xr.DataArray,
-    gain: float,
-    w: Window,
-    w_pad: Window,
-) -> None:
-    x = darray.data.compute(num_workers=1)
-
-    image_height = window_size + padding * 2
-    image_width = window_size + padding * 2
-
-    # Get row adjustments
-    row_before_padded = abs(w_pad.row_off - w.row_off)
-    row_before_to_pad = padding - row_before_padded
-    row_after_to_pad = image_height - w_pad.height - row_before_to_pad
-
-    # Get column adjustments
-    col_before_padded = abs(w_pad.col_off - w.col_off)
-    col_before_to_pad = padding - col_before_padded
-    col_after_to_pad = image_width - w_pad.width - col_before_to_pad
-
-    x = np.pad(
-        x,
-        pad_width=(
-            (0, 0),
-            (0, 0),
-            (row_before_to_pad, row_after_to_pad),
-            (col_before_to_pad, col_after_to_pad),
-        ),
-        mode="constant",
-        constant_values=0,
-    )
-
-    x = einops.rearrange(
-        torch.from_numpy(x / gain).to(dtype=torch.int32),
-        't c h w -> 1 c t h w',
-    )
-
-    assert x.shape[-2:] == (
-        image_height,
-        image_width,
-    ), "The padded array does not have the correct height/width dimensions."
-
-    batch_id = f"{region}_{start_date}_{end_date}_{w.row_off}_{w.col_off}"
-
-    batch = Data(
-        x=x,
-        start_year=[start_date],
-        end_year=[end_date],
-        padding=[padding],
-        window_row_off=[w.row_off],
-        window_col_off=[w.col_off],
-        window_height=[w.height],
-        window_width=[w.width],
-        window_pad_row_off=[w_pad.row_off],
-        window_pad_col_off=[w_pad.col_off],
-        window_pad_height=[w_pad.height],
-        window_pad_width=[w_pad.width],
-        row_before_to_pad=[row_before_to_pad],
-        row_after_to_pad=[row_after_to_pad],
-        col_before_to_pad=[col_before_to_pad],
-        col_after_to_pad=[col_after_to_pad],
-        res=[res],
-        resampling=[resampling],
-        left=[darray.gw.left],
-        bottom=[darray.gw.bottom],
-        right=[darray.gw.right],
-        top=[darray.gw.top],
-        batch_id=[batch_id],
-    )
-
-    batch.to_file(
-        write_path / f"{batch_id}.pt",
-        compress=compress_method,
-    )
-
-
-@threadpool_limits.wrap(limits=1, user_api="blas")
-def create_predict_dataset(
-    image_list: T.List[T.List[T.Union[str, Path]]],
-    region: str,
-    process_path: Path = None,
-    date_format: str = "%Y%j",
-    gain: float = 1e-4,
-    offset: float = 0.0,
-    ref_res: T.Union[float, T.Tuple[float, float]] = 10.0,
-    resampling: str = "nearest",
-    window_size: int = 100,
-    padding: int = 101,
-    num_workers: int = 1,
-    chunksize: int = 100,
-    compress_method: T.Union[int, str] = 'zlib',
-):
-    with gw.config.update(ref_res=ref_res):
-        with gw.open(
-            image_list,
-            stack_dim="band",
-            band_names=list(range(1, len(image_list) + 1)),
-            resampling=resampling,
-            chunks=512,
-        ) as src_ts:
-
-            windows = get_window_offsets(
-                src_ts.gw.nrows,
-                src_ts.gw.ncols,
-                window_size,
-                window_size,
-                padding=(padding, padding, padding, padding),
-            )
-
-            num_time, num_bands = get_image_list_dims(image_list, src_ts)
-
-            time_series: xr.DataArray = reshape_and_mask_array(
-                data=src_ts,
-                num_time=num_time,
-                num_bands=num_bands,
-                gain=gain,
-                offset=offset,
-            )
-
-            partial_create_and_save_window = partial(
-                create_and_save_window,
-                write_path=process_path,
-                res=ref_res,
-                resampling=resampling,
-                region=region,
-                start_date=pd.to_datetime(
-                    Path(image_list[0]).stem, format=date_format
-                ).strftime("%Y%m%d"),
-                end_date=pd.to_datetime(
-                    Path(image_list[-1]).stem, format=date_format
-                ).strftime("%Y%m%d"),
-                window_size=window_size,
-                padding=padding,
-                compress_method=compress_method,
-                gain=gain,
-            )
-
-            with parallel_backend(backend="threading", n_jobs=num_workers):
-                for window_chunk in get_window_chunk(windows, chunksize):
-                    with Parallel(temp_folder="/tmp") as pool:
-                        __ = pool(
-                            delayed(partial_create_and_save_window)(
-                                darray=read_slice(time_series, window_pad),
-                                w=window,
-                                w_pad=window_pad,
-                            )
-                            for window, window_pad in window_chunk
-                        )
-
-
-def get_reference_bounds(
-    df_grid: gpd.GeoDataFrame,
-    grid_size: tuple,
-    filename: T.Union[Path, str],
-    ref_res: tuple,
-) -> T.List[float]:
-    ref_bounds = df_grid.total_bounds.tolist()
-
-    if grid_size is not None:
-        # Enforce bounds given height/width dimensions
-        height, width = grid_size
-        left, bottom, right, top = ref_bounds
-
-        with gw.open(filename) as src:
-            image_crs = src.gw.crs_to_pyproj
-            if ref_res is None:
-                ref_res = (src.gw.celly, src.gw.cellx)
-            else:
-                ref_res = (ref_res, ref_res)
-
-        (dst_transform, dst_width, dst_height,) = calculate_default_transform(
-            src_crs=image_crs,
-            dst_crs=df_grid.crs,
-            width=int(abs(round((right - left) / ref_res[1]))),
-            height=int(abs(round((top - bottom) / ref_res[0]))),
-            left=left,
-            bottom=bottom,
-            right=right,
-            top=top,
-            dst_width=width,
-            dst_height=height,
-        )
-        dst_left = dst_transform[2]
-        dst_top = dst_transform[5]
-        dst_right = dst_left + abs(dst_width * dst_transform[0])
-        dst_bottom = dst_top - abs(dst_height * dst_transform[4])
-        ref_bounds = [dst_left, dst_bottom, dst_right, dst_top]
-
-    return ref_bounds
 
 
 @threadpool_limits.wrap(limits=1, user_api="blas")
